+++ lib/buzzn/schemas/constraints/reading/single.rb
@@ -1,8 +1,8 @@
 require_relative '../reading'
 
 Schemas::Constraints::Reading::Single = Schemas::Support.Form do
-  required(:raw_value).filled(:float?)
-  required(:value).filled(:float?)
+  required(:raw_value).filled(:int?)
+  required(:value).filled(:int?)
   required(:unit).value(included_in?: Reading::Single.units.values)
   required(:reason).value(included_in?: Reading::Single.reasons.values)
   required(:read_by).value(included_in?: Reading::Single.read_by.values)
